Add `./scripts/freeze` to freeze/format all dependencies
... and run the script!

The script uses the `--cache` flag to `dhall freeze` so that
imports gracefully fallback if there is disruption to a semantic
integrity check.

This script is essentially the same thing that the Prelude does.

The script also freezes internal imports so that users can efficiently
import the top-level `package.dhall` import unqualified and internal
references are still automatically cached on their behalf.